refactor: 게임과 스크린샷 스크래퍼를 분리함
