+++ data_scrapers/tests/public/scraper/test_screenshot_scraper.py
@@ -1,6 +1,6 @@
 import pytest
 
-from public.scraper.service import scrap_game_screenshot, scrap_games
+from public.scraper.screenshot import scrap_game_screenshot
 from tests.public.utils.mock_lambda_api import MockLambdaAPI
 from tests.public.utils.mock_steam_api import MockSteamAPI
 from tests.public.utils.model import create_random_game
@@ -9,29 +9,6 @@ from tests.public.utils.model import create_random_game
 @pytest.fixture
 def mock_steam_api() -> MockSteamAPI:
     return MockSteamAPI()
-
-
-def test_scrap_geams는_새_게임을_저장한다(mock_steam_api: MockSteamAPI):
-    mock_steam_api.prepare_mock_data()
-    lambda_api = MockLambdaAPI()
-
-    scrap_games(mock_steam_api, lambda_api)
-
-    assert len(lambda_api.games.values()) > 0
-
-
-def test_scrap_games는_중복된_게임은_다시_저장하지_않는다(mock_steam_api: MockSteamAPI):
-    mock_steam_api.prepare_mock_data()
-    lambda_api = MockLambdaAPI()
-
-    scrap_games(mock_steam_api, lambda_api)
-    before_scraped = lambda_api.games.copy()
-
-    # duplicated games
-    scrap_games(mock_steam_api, lambda_api)
-    after_scraped = lambda_api.games
-
-    assert before_scraped == after_scraped
 
 
 def test_scrap_screenshots은_스크린샷을_저장한다(mock_steam_api: MockSteamAPI):
